Update table type
Table type update

+++ weather/database/migrations/2019_11_27_100236_create_feedback_table.php
@@ -14,11 +14,11 @@ class CreateFeedbackTable extends Migration
     public function up()
     {
         Schema::create('feedback', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name')->nullable();;
             $table->string('email')->nullable();;
             $table->string('text');
-            $table->integer('user_id')->nullable();
+            $table->unsignedBigInteger('user_id')->nullable();
             $table->timestamps();
         });
     }
